Fix tests into working condition
Add rule to disallow registering to an unpublished event
This was possible if registration was open but the event was
unpublished and you manually make the post request

Change the date-test command so that you can provide an optional argument
that specifies which test module to run. Ex. `date-test news` will run only
tests found in the "news" module.