Fix tests into working condition (#376)
* Fix tests into working condition

Fix all tests so that they don't give random errors and so that they pass for the current state of origin/develop

Add unit tests to CI pipeline

Change the date-test command so that you can provide an optional argument
that specifies which test module to run. Ex. `date-test news` will run only
tests found in the "news" module.

* Add env variable that gets set for tests only

* Use tempfile module to simplify document testing

* Add testing to CI-pipeline

* CI hotfix

* CI hotfix #2

* Final cleanup for tests PR